requirements.txt
numpy
face_recognition
opencv-python
tensorflow
keras
Pillow
tqdm